infra: moved regex to utils/
